working server

+++ routes/todo.js
@@ -1,11 +1,17 @@
 import { Router } from "express";
+import { verifyAuth } from "./helper";
 import db from "../models";
 
 const todoRoutes = Router();
 
-todoRoutes.get("/api/todos", (_, res) => {
+todoRoutes.get("/api/todos", verifyAuth, (req, res) => {
+  const currentUser = req.user;
   return db.todo
-    .findAll()
+    .findAll({
+      where: {
+        user_id: currentUser.id,
+      },
+    })
     .then((todo) => res.send(todo))
     .catch((err) => {
       console.log("error occurred", JSON.stringify(err));
@@ -13,18 +19,21 @@ todoRoutes.get("/api/todos", (_, res) => {
     });
 });
 
-todoRoutes.post("/api/todos", (req, res) => {
+todoRoutes.post("/api/todos", verifyAuth, (req, res) => {
+  const currentUser = req.user;
+
   const { name, description } = req.body;
 
   return db.todo
-    .create({ name, description })
+    .create({ name, description, user_id: currentUser.id })
     .then(() => res.send(todo))
     .catch((err) => res.send(err));
 });
 
-todoRoutes.put("/api/todos/:id", (req, res) => {
+todoRoutes.put("/api/todos/:id", verifyAuth, (req, res) => {
   const id = parseInt(req.params.id);
   const { name, description } = req.body;
+  const currentUser = req.user;
 
   const payload = {
     name,
@@ -33,17 +42,18 @@ todoRoutes.put("/api/todos/:id", (req, res) => {
 
   return db.todo
     .update(payload, {
-      where: { id },
+      where: { id, user_id: currentUser.id },
     })
     .then((todo) => res.send(todo))
     .catch((err) => res.send(err));
 });
 
-todoRoutes.delete("/api/todos/:id", (req, res) => {
+todoRoutes.delete("/api/todos/:id", verifyAuth, (req, res) => {
   const id = parseInt(req.params.id);
+  const currentUser = req.user;
 
   return db.todo
-    .destroy({ where: { id } })
+    .destroy({ where: { id, user_id: currentUser.id } })
     .then(() => res.send({}))
     .catch((err) => res.send(err));
 });
